new compress() end memory_footprint_bytes() routines to create compressed matrices with external pointers; matrices can be new created on compressed raw memory; fixed a bug in the offset() computation for lower triangular matrices: num_elements_c was assumed to be !=0 before

+++ src/include/general_matrix_vector_kernel.hpp
@@ -11,8 +11,8 @@
 #include <fp/fp_blas.hpp>
 
 // fundamental real data type: 'float' or 'double'
-using real_t = float;
-//using real_t = double;
+//using real_t = float;
+using real_t = double;
 
 // number of bits to be used for exponent and mantissa
 #if defined(_BE)
